Diet of mice for gene expression data

mouse_diet contains the mouse ID and diet information for the mice used in this study.

moexpr_data

mouse_diet

Format

moexpr_data is a tibble (18,596 x 49)

gene

gene name for the row

GSM994790 ... GSM994837

expression data for each mouse

#' @format mouse_diet is a tibble (48 x 2)

mouse

the ID of the mouse

diet

the level of fat in the diet

Source

https://www.ncbi.nlm.nih.gov/sites/GDSbrowser/

Details

moexpr_data contains the gene expression of each mouse (one mouse per column)

This data was obtained from the Gene Expression Omnibus (GEO) under DataSet Record GDS6248.

"Analysis of livers of C57BL/6J mice fed a high fat diet for up to 24 weeks. Significant body weight gain was observed after 4 weeks. Results provide insight into the effect of high fat diets on metabolism in the liver."

Examples

mouse_diet

#> # A tibble: 48 x 2
#>    mouse_id  diet  
#>    <chr>     <chr> 
#>  1 GSM994790 Normal
#>  2 GSM994791 Normal
#>  3 GSM994792 Normal
#>  4 GSM994793 Normal
#>  5 GSM994794 Normal
#>  6 GSM994795 Normal
#>  7 GSM994796 Normal
#>  8 GSM994797 Normal
#>  9 GSM994798 Normal
#> 10 GSM994799 Normal
#> # … with 38 more rows

moexpr_data

#> # A tibble: 18,596 x 49
#>    gene  GSM994790 GSM994791 GSM994792 GSM994793 GSM994794 GSM994795 GSM994796
#>    <chr>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 Aaas       6.84      6.84      6.76      6.70      6.71      6.74      6.65
#>  2 Aacs       8.82      8.55      9.06     10.5      10.7       9.26      9.73
#>  3 Aadac      8.29      8.71      8.48      8.58      8.33      8.35      9.23
#>  4 Aada…      7.04      6.89      6.94      6.71      6.71      6.83      6.79
#>  5 Aadat      8.47      8.33      8.37      8.21      8.21      8.13      8.79
#>  6 Aaed1      8.58      8.63      8.71      8.20      8.55      8.34      8.79
#>  7 Aagab      9.93      9.93     10.1       9.96     10.1      10.0       9.81
#>  8 Aak1       7.03      7.26      7.25      7.05      7.15      7.24      7.10
#>  9 Aamdc      9.57      9.55     10.0       9.89     10.3      10.3       9.86
#> 10 Aamp      11.7      11.8      11.7      11.8      11.7      11.7      11.7 
#> # … with 18,586 more rows, and 41 more variables: GSM994797 <dbl>,
#> #   GSM994798 <dbl>, GSM994799 <dbl>, GSM994800 <dbl>, GSM994801 <dbl>,
#> #   GSM994802 <dbl>, GSM994803 <dbl>, GSM994804 <dbl>, GSM994805 <dbl>,
#> #   GSM994806 <dbl>, GSM994807 <dbl>, GSM994808 <dbl>, GSM994809 <dbl>,
#> #   GSM994810 <dbl>, GSM994811 <dbl>, GSM994812 <dbl>, GSM994813 <dbl>,
#> #   GSM994814 <dbl>, GSM994815 <dbl>, GSM994816 <dbl>, GSM994817 <dbl>,
#> #   GSM994818 <dbl>, GSM994819 <dbl>, GSM994820 <dbl>, GSM994821 <dbl>,
#> #   GSM994822 <dbl>, GSM994823 <dbl>, GSM994824 <dbl>, GSM994825 <dbl>,
#> #   GSM994826 <dbl>, GSM994827 <dbl>, GSM994828 <dbl>, GSM994829 <dbl>,
#> #   GSM994830 <dbl>, GSM994831 <dbl>, GSM994832 <dbl>, GSM994833 <dbl>,
#> #   GSM994834 <dbl>, GSM994835 <dbl>, GSM994836 <dbl>, GSM994837 <dbl>