Gradient colour scales for bottom-right triangle in geom_asymmat

This dictates a gradient color scheme for the bottom-right (br) of a geom_asymmat ggplot geom. scale_*_br_gradient creates a two colour gradient (low-high), scale_*_br_gradient2 creates a diverging colour gradient (low-mid-high), scale_*_br_gradientn creates a n-colour gradient.

scale_fill_br_gradient(..., low = "#132B43", high = "#56B1F7",
  space = "Lab", na.value = "grey50", guide = "colourbar",
  aesthetics = "fill_br")

Arguments

...
arguments passed on to continuous_scale_asym

low
the colors to represent low and high values

high
the colors to represent low and high values

space
colour space in which to calculate gradient. Must be "Lab" - other values are deprecated.

na.value
color of missing (NA) values

guide
Type of legend. Use "colourbar" for continuous colour bar, or "legend" for discrete colour legend.

aesthetics
Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. For now, leave the default alone, though I plan to reinstate the standard ggplot system here, eventually.

Examples

# vignette(ggasym)